--- chapter-12/src/11-replace-superclass-with-delegate/CatalogItem-01.js
@@ -1,19 +1,0 @@
-class CatalogItem {
-  constructor(id, title, tags) {
-    this._id = id;
-    this._title = title;
-    this._tags = tags;
-  }
-
-  get id() {
-    return this._id;
-  }
-
-  get title() {
-    return this._title;
-  }
-
-  hasTag(arg) {
-    return this._tags.includes(arg);
-  }
-}
