+++ package.json
@@ -22,6 +22,7 @@
   },
   "repository": "https://github.com/pocojang/refactoring-2nd-edition.git",
   "dependencies": {
+    "@js-joda/core": "^5.5.3",
     "lodash": "^4.17.21"
   }
 }
